requirements.txt
openai==0.28.1
pandas==2.1.2
pandasql==0.7.3
python-dotenv==1.0.0